backend/requirements.txt
# ============================================================
# PetroVision Backend — Python Dependencies
# UTN FRM — Tesis ISI 2026
# ============================================================

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database — Relational (PostgreSQL)
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Database — Time Series (InfluxDB)
influxdb-client==1.40.0

# Cache / Message Queue
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
xgboost==2.0.3
openpyxl>=3.1.0

# WebSocket support (included in uvicorn[standard])
websockets==12.0

# Web Push Notifications (VAPID)
pywebpush==2.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
